feat: move data.ts to respective files
